requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-jinja2==1.5
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.1
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comet-ml==3.31.19
configobj==5.0.6
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
dnspython==2.2.1
docker-pycreds==0.4.0
docstring-parser==0.15
dulwich==0.20.50
einops==0.6.0
email-validator==1.3.0
everett==3.1.0
fastapi==0.68.1
ffmpy==0.3.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
gradio==3.12.0
grpcio==1.51.1
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.1
idna==3.4
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.1.0
importlib-resources==5.10.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonargparse==4.18.0
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.8
linkify-it-py==1.0.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.3
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.5.3.56
opencv-python-headless==4.6.0.66
orjson==3.8.3
packaging==22.0
pandas==1.5.2
paramiko==2.12.0
pathtools==0.1.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
promise==2.3
protobuf==3.20.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pyDeprecate==0.3.2
pydub==0.25.1
PyNaCl==1.5.0
pyngrok==5.2.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytorch-lightning==1.6.5
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
semantic-version==2.10.0
sentry-sdk==1.11.1
setproctitle==1.3.2
Shapely==1.8.5.post1
shortuuid==1.0.11
simplejson==3.18.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.14.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.5.4
torch==1.12.0
torchmetrics==0.9.3
torchtext==0.13.0
torchvision==0.13.0
tqdm==4.64.1
typing_extensions==4.4.0
uc-micro-py==1.0.1
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
wandb==0.12.17
watchfiles==0.18.1
websocket-client==1.3.3
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
wurlitzer==3.0.3
yarl==1.8.2
zipp==3.11.0
